requirements/base.txt
Django==2.2.1
django-cors-headers==3.0.1
django-model-utils==3.1.2
djangorestframework==3.9.4
Pillow==6.0.0
psycopg2==2.8.2
arrow==0.13.2
elasticsearch==6.3.1
elasticsearch-dsl==6.4.0
django-rest-elasticsearch==0.4.1
drf-yasg==1.15.0
swagger-spec-validator==2.4.3
flex==6.14.0
django-filter==2.1.0